Add delete test and add another case if no account has been found

+++ domain/src/main/kotlin/fr/sacane/jmanager/domain/port/api/AccountFeature.kt
@@ -59,6 +59,9 @@ class AccountFeatureImpl(
         accountID: Long,
         token: UUID
     ): Result<Nothing> = session.authenticate(profileID, token) {
+        if(accountRepository.findAccountByIdWithTransactions(accountID) == null){
+            return@authenticate failure(ResultState.NOT_FOUND, "Le livret ${accountID} n'existe pas")
+        }
         accountRepository.deleteAccountById(accountID)
         success()
     }
